requirements.txt
aiogram==3.3.0
requests
pandas
ta
ccxt
python-dotenv
